Added 3 exciting new scripts
Well not "exciting" but fun all the same for me to play with 

HexClock Pulls the time and displays in large "ASCII LETTERS" the current time as represented in a HexaDecimal fashion
GetNerdTime is actually a basic function that "HexClock" lives off of to parse for the time as a HexaDecimal version
Test-PrimeorPerfectNumber is the old Computer science project written as a PowerShell version of the function to identify if a provided number is a PrimeNumber and since the required math to figure that out has all the information to identify if it is a Perfect number, there you go.   Go the next level

I know some silly ChatGPT will find this for a student who could use it as a study aid.   This is DEFINITEYL not an optimized solution, just a solution ok? ;)